requirements-ci.txt
coverage
unittest-xml-reporting
python-jenkins
xmltodict